#patch Cambio en el saldo de Double/Long/Integer a Long

+++ source/eburyApp/eburyApp-ejb/src/main/java/es/uma/softcoders/eburyApp/ejb/GestionTransaccion.java
@@ -26,5 +26,5 @@ public interface GestionTransaccion {
 	 * @throws DivisaInexistenteException 
 	 * @author Marta Maleno Escudero
 	 *  */
-	public void cambioDivisa(String cuentaPool, String divOrigen, String divDestino, Double cantidad) throws CuentaNoExistenteException, DivisaInexistenteException;
+	public void cambioDivisa(String cuentaPool, String divOrigen, String divDestino, Long cantidad) throws CuentaNoExistenteException, DivisaInexistenteException;
 }
